requirements.txt
asgiref==3.4.1
certifi==2021.5.30
click==8.0.1
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.4
Django==3.2.6
django-cors-headers==3.8.0
django-geojson==3.2.0
djangorestframework==3.12.4
geohash-hilbert==1.4.0
geojson==2.5.0
geopip==1.1
importlib-metadata==4.6.4
jsonfield==3.1.0
munch==2.5.0
numpy==1.19.5
pandas==1.1.5
pyproj==3.0.1
python-dateutil==2.8.2
pytz==2021.1
Shapely==1.7.1
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
zipp==3.5.0
